+++ layer/tests/tests.cpp
@@ -756,7 +756,7 @@ TEST_F(LayerTests, TestQueueFamilyPropertiesPartial) {
                 if ((device_qf_props[i].queueFlags & VK_QUEUE_COMPUTE_BIT) == VK_QUEUE_COMPUTE_BIT &&
                     device_qf_props[i].minImageTransferGranularity.width <= 4 &&
                     device_qf_props[i].minImageTransferGranularity.height <= 4) {
-                    device_queue_index = i;
+                    device_queue_index = static_cast<int32_t>(i);
                     break;
                 }
             }
@@ -768,8 +768,8 @@ TEST_F(LayerTests, TestQueueFamilyPropertiesPartial) {
             ASSERT_EQ(qf_props[0].minImageTransferGranularity.height, 4u);
 
             if (device_queue_index != -1) {
-                ASSERT_EQ(qf_props[0].minImageTransferGranularity.depth, device_qf_props[device_queue_index].minImageTransferGranularity.depth);
-                ASSERT_EQ(qf_props[0].timestampValidBits, device_qf_props[device_queue_index].timestampValidBits);
+                ASSERT_EQ(qf_props[0].minImageTransferGranularity.depth, device_qf_props[static_cast<std::size_t>(device_queue_index)].minImageTransferGranularity.depth);
+                ASSERT_EQ(qf_props[0].timestampValidBits, device_qf_props[static_cast<std::size_t>(device_queue_index)].timestampValidBits);
             }
         }
     }
